Add NotNullAttribute support
