+++ Editor/Core/UIElements/Graph/Ports/CeresEdge.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 7c71bdf80aa02fb478259289c2e073fa
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
