+++ Editor/Core/UIElements/Graph/Ports/CeresPortView.cs
@@ -1,5 +1,6 @@
 using System;
 using System.Collections.Generic;
+using System.Diagnostics.CodeAnalysis;
 using System.Linq;
 using System.Reflection;
 using Ceres.Annotations;
@@ -17,99 +18,6 @@ using UObject = UnityEngine.Object;
 
 namespace Ceres.Editor.Graph
 {
-    public class CeresEdge : Edge
-    {
-        public CeresEdge()
-        {
-            AddToClassList(nameof(CeresEdge));
-        }
-    }
-    
-    /// <summary>
-    /// Edge listener connecting <see cref="CeresEdge"/> between <see cref="CeresPortElement"/>
-    /// </summary>
-    public class CeresEdgeListener : IEdgeConnectorListener
-    {
-        private readonly GraphViewChange _graphViewChange;
-        
-        private readonly List<Edge> _edgesToCreate;
-        
-        private readonly List<GraphElement> _edgesToDelete;
-        
-        public CeresGraphView GraphView { get; }
-        
-        public CeresEdgeListener(CeresGraphView ceresGraphView)
-        {
-            GraphView = ceresGraphView;
-            _edgesToCreate = new List<Edge>();
-            _edgesToDelete = new List<GraphElement>();
-            _graphViewChange.edgesToCreate = _edgesToCreate;
-        }
-        
-        public void OnDropOutsidePort(Edge edge, Vector2 position)
-        {
-            var screenPosition = GUIUtility.GUIToScreenPoint(
-                Event.current.mousePosition
-            );
-            
-            if (edge.output?.edgeConnector.edgeDragHelper.draggedPort != null)
-            {
-                GraphView.OpenSearch(
-                    screenPosition,
-                    ((CeresPortElement)edge.output.edgeConnector.edgeDragHelper.draggedPort).View
-                );
-            }
-            else if (edge.input?.edgeConnector.edgeDragHelper.draggedPort != null)
-            {
-                GraphView.OpenSearch(
-                    screenPosition,
-                    ((CeresPortElement)edge.input.edgeConnector.edgeDragHelper.draggedPort).View
-                );
-            }
-        }
-
-        public void OnDrop(GraphView graphView, Edge edge)
-        {
-            _edgesToCreate.Clear();
-            _edgesToCreate.Add(edge);
-            _edgesToDelete.Clear();
-            
-            if (edge.input.capacity == Port.Capacity.Single)
-            {
-                foreach (var connection in edge.input.connections)
-                {
-                    if (connection != edge)
-                        _edgesToDelete.Add(connection);
-                }
-            }
-            
-            if (edge.output.capacity == Port.Capacity.Single)
-            {
-                foreach (var connection in edge.output.connections)
-                {
-                    if (connection != edge)
-                        _edgesToDelete.Add(connection);
-                }
-            }
-            
-            if (_edgesToDelete.Count > 0)
-            {
-                graphView.DeleteElements(_edgesToDelete);
-            }
-            
-            var edgesToCreate = _edgesToCreate;
-            if (graphView.graphViewChanged != null)
-            {
-                edgesToCreate = graphView.graphViewChanged(_graphViewChange).edgesToCreate;
-            }
-            foreach (var edgeToCreate in edgesToCreate)
-            {
-                graphView.AddElement(edgeToCreate);
-                edgeToCreate.input.Connect(edgeToCreate);
-                edgeToCreate.output.Connect(edgeToCreate);
-            }
-        }
-    }
     public class CeresPortElement: Port
     {
         public CeresPortView View { get; private set; }
@@ -365,9 +273,13 @@ namespace Ceres.Editor.Graph
 
         public bool AlwaysConnected()
         {
+            if (BindingType == PortBindingType.Parameter)
+            {
+                var attribute = ResolvedParameterInfo.GetCustomAttribute<NotNullAttribute>();
+                return attribute != null;
+            }
             var inputPort = PortFieldInfo.GetCustomAttribute<InputPortAttribute>();
-            if (inputPort == null) return false;
-            return inputPort.AlwaysConnected;
+            return inputPort is { AlwaysConnected: true };
         }
 
         public Port.Capacity GetCapacity()
@@ -623,7 +535,7 @@ namespace Ceres.Editor.Graph
         /// <param name="parameterInfo"></param>
         public void SetDisplayDataFromParameterInfo(ParameterInfo parameterInfo)
         {
-            if(Binding.BindingType == CeresPortViewBinding.PortBindingType.Field)
+            if (Binding.BindingType == CeresPortViewBinding.PortBindingType.Field)
             {
                 /* Set default value if exist */
                 if (FieldResolver != null && parameterInfo.HasDefaultValue)
@@ -633,6 +545,10 @@ namespace Ceres.Editor.Graph
                 SetDisplayName(parameterInfo.Name);
                 var fieldShowInEditor = parameterInfo.GetCustomAttribute<HideInGraphEditorAttribute>() == null;
                 PortElement.SetEditorFieldVisibility(fieldShowInEditor);
+                if (parameterInfo.GetCustomAttribute<NotNullAttribute>() != null)
+                {
+                    Flags |= CeresPortViewFlags.ValidateConnection;
+                }
             }
             else
             {
